+++ probeye/definition/inverse_problem.py
@@ -60,8 +60,11 @@ class InverseProblem:
         print_header: bool = True,
     ):
 
-        # the name of the problem
+        # the name and safe name of the problem; the latter is intended for file naming
         self.name = name
+        self.safe_name = safe_string(name)
+
+        # logging related preparations
         self.use_default_logger = use_default_logger
         self.log_level = log_level
         self.log_file = log_file