+++ probeye/ontology/__init__.py
@@ -1,0 +1,3 @@
+# local imports
+from probeye.ontology import knowledge_graph_export
+from probeye.ontology import knowledge_graph_import