+++ tests/integration_tests/tests_without_correlation/test_query_results.py
@@ -1,12 +1,13 @@
 """
-                                   Linear regression
+                     Linear regression (with knowledge graph query)
 ----------------------------------------------------------------------------------------
-                    ---> Multiplicative model prediction error <---
+                       ---> Additive model prediction error <---
 ----------------------------------------------------------------------------------------
 The model equation is y(x) = a * x + b with a, b being the model parameters, while the
-likelihood model is based on a normal unit-mean multiplicative model error distribution
-with the standard deviation to infer and an additive measurement error distribution with
-known standard deviation. The problem is approached via max likelihood estimation.
+likelihood model is based on a normal zero-mean additive model error distribution with
+the standard deviation to infer. The problem is approached via emcee. After the problem
+and the sampling results have been exported to a knowledge graph, this graph is queried
+to obtain the problem's parameters and the paths to the files that contain the samples.
 """
 
 # standard library imports
@@ -25,23 +26,21 @@ from probeye.definition.likelihood_model import GaussianLikelihoodModel
 
 # local imports (knowledge graph)
 from probeye.ontology.knowledge_graph_export import export_knowledge_graph
+from probeye.ontology.knowledge_graph_import import import_parameter_samples
+from probeye.ontology.knowledge_graph_export import export_results_to_knowledge_graph
 
 # local imports (testing related)
-from tests.integration_tests.subroutines import run_inference_engines
+from probeye.inference.emcee.solver import EmceeSolver
 
 
 class TestProblem(unittest.TestCase):
-    def test_linear_regression_multiplicative_model_error(
+    def test_query_results(
         self,
-        n_steps: int = 200,
-        n_initial_steps: int = 100,
-        n_walkers: int = 20,
+        n_steps: int = 100,
+        n_initial_steps: int = 50,
+        n_walkers: int = 10,
         plot: bool = False,
         show_progress: bool = False,
-        write_to_graph: bool = True,
-        run_scipy: bool = True,
-        run_emcee: bool = False,  # intentionally False for faster test-runs
-        run_dynesty: bool = False,  # intentionally False for faster test-runs
     ):
         """
         Integration test for the problem described at the top of this file.
@@ -61,17 +60,6 @@ class TestProblem(unittest.TestCase):
             plots are closed.
         show_progress
             If True, progress-bars will be shown, if available.
-        write_to_graph
-            Triggers the export of the solver results to a given knowledge graph.
-        run_scipy
-            If True, the problem is solved with scipy (maximum likelihood est).
-            Otherwise, no maximum likelihood estimate is derived.
-        run_emcee
-            If True, the problem is solved with the emcee solver. Otherwise, the emcee
-            solver will not be used.
-        run_dynesty
-            If True, the problem is solved with the dynesty solver. Otherwise, the
-            dynesty solver will not be used.
         """
 
         # ============================================================================ #
@@ -88,13 +76,10 @@ class TestProblem(unittest.TestCase):
         mean_b = 1.0
         std_b = 1.0
 
-        # 'true' value of multiplicative error sd, and its uniform prior parameters
-        sigma = 0.1
+        # 'true' value of additive error sd, and its uniform prior parameters
+        sigma = 0.5
         low_sigma = 0.0
-        high_sigma = 0.5
-
-        # assuming a known measurement error
-        sigma_m = 0.02
+        high_sigma = 0.8
 
         # the number of generated experiment_names and seed for random numbers
         n_tests = 50
@@ -108,9 +93,7 @@ class TestProblem(unittest.TestCase):
             def interface(self):
                 self.parameters = [{"a": "m"}, "b"]
                 self.input_sensors = Sensor("x")
-                self.output_sensors = Sensor(
-                    "y", std_model="sigma", std_measurement="sigma_m"
-                )
+                self.output_sensors = Sensor("y", std_model="sigma")
 
             def response(self, inp: dict) -> dict:
                 x = inp["x"]
@@ -123,7 +106,7 @@ class TestProblem(unittest.TestCase):
         # ============================================================================ #
 
         # initialize the inverse problem with a useful name
-        problem = InverseProblem("Linear regression (MME)")
+        problem = InverseProblem("Linear regression (AME)")
 
         # add all parameters to the problem
         problem.add_parameter(
@@ -145,15 +128,8 @@ class TestProblem(unittest.TestCase):
             "likelihood",
             domain="(0, +oo)",
             tex=r"$\sigma$",
-            info="Standard deviation, of unit-mean multiplicative model error",
+            info="Standard deviation, of zero-mean additive model error",
             prior=("uniform", {"low": low_sigma, "high": high_sigma}),
-        )
-        problem.add_parameter(
-            "sigma_m",
-            "likelihood",
-            tex=r"$\sigma_m$",
-            info="Standard deviation, of zero-mean additive measurement error",
-            const=sigma_m,
         )
 
         # add the forward model to the problem
@@ -170,7 +146,7 @@ class TestProblem(unittest.TestCase):
         y_true = linear_model.response(
             {linear_model.input_sensor.name: x_test, "m": a_true, "b": b_true}
         )[linear_model.output_sensor.name]
-        y_test = np.random.normal(loc=y_true, scale=sigma * y_true + sigma_m)
+        y_test = np.random.normal(loc=y_true, scale=sigma)
 
         # add the experimental data
         problem.add_experiment(
@@ -199,10 +175,10 @@ class TestProblem(unittest.TestCase):
         # add the likelihood model to the problem
         problem.add_likelihood_model(
             GaussianLikelihoodModel(
-                prms_def=["sigma", "sigma_m"],
+                prms_def="sigma",
                 experiment_name="TestSeries_1",
-                model_error="multiplicative",
-                additive_measurement_error=True,
+                model_error="additive",
+                name="SimpleLikelihoodModel",
             )
         )
 
@@ -220,27 +196,39 @@ class TestProblem(unittest.TestCase):
         export_knowledge_graph(problem, knowledge_graph_file, data_dir=dir_path)
 
         # ============================================================================ #
-        #                    Solve problem with inference engine(s)                    #
+        #        Solve problem with inference engine and write results to graph        #
         # ============================================================================ #
 
-        # this routine is imported from another script because it it used by all
-        # integration tests in the same way; ref_values are used for plotting
+        # run inference step using emcee
         true_values = {"a": a_true, "b": b_true, "sigma": sigma}
-        run_inference_engines(
-            problem,
-            true_values=true_values,
+        emcee_solver = EmceeSolver(problem, show_progress=show_progress)
+        inference_data = emcee_solver.run_mcmc(
+            n_walkers=n_walkers,
             n_steps=n_steps,
             n_initial_steps=n_initial_steps,
-            n_walkers=n_walkers,
-            plot=plot,
-            show_progress=show_progress,
-            write_to_graph=write_to_graph,
-            knowledge_graph_file=knowledge_graph_file,
-            data_dir=dir_path,
-            run_scipy=run_scipy,
-            run_emcee=run_emcee,
-            run_dynesty=run_dynesty,
+            true_values=true_values,
         )
+
+        # export the results from the 'inference_data' object to the graph
+        export_results_to_knowledge_graph(
+            problem,
+            inference_data,
+            knowledge_graph_file,
+            data_dir=dir_path,
+        )
+
+        # ============================================================================ #
+        #                          Query the knowledge graph                           #
+        # ============================================================================ #
+
+        # get the samples from the knowledge graph
+        sample_dict = import_parameter_samples(knowledge_graph_file)
+
+        # check if the parameters returned from the query are the same as the ones
+        # defined within the problem scope
+        prm_names_from_query = set(sample_dict.keys())
+        prm_names_in_problem = set(problem.latent_prms)
+        self.assertTrue(prm_names_from_query == prm_names_in_problem)
 
 
 if __name__ == "__main__":